s...stable?
