+++ scripts/grid.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+java -jar ./selenium/selenium-server-standalone-3.7.1.jar \
+    -role hub
