+++ src/__init__.py
@@ -1,4 +1,6 @@
 from .env_conf import EnvKeyNotFound, EnvValueUnexpected, get_credentials
 from driver_selector import get_webdriver
-from page_objects import PageObject, FiltersApplyingPageObject, LOGIN_PAGE_URL, SECRET_PAGE_URL, InitPage, HeaderObject
+from page_objects import PageObject, FiltersApplyingPageObject, LOGIN_PAGE_URL, SECRET_PAGE_URL
 from .message_activities import MessageActivities
+
+import ax
