+++ src/ax/__init__.py
@@ -1,0 +1,4 @@
+from base_page_object import BasePageObject
+from header_object import HeaderObject
+from init_page import InitPage
+from consts import LOGIN_PAGE, LOGIN_PAGE_URL, SECRET_PAGE_URL
