+++ src/ax/consts.py
@@ -1,0 +1,9 @@
+
+LOGIN_PAGE = "LOGIN_PAGE-state"
+LOGIN_PAGE_URL = "https://account.mail.ru/login?success_redirect=https://octavius.mail.ru&allow_external=1"
+SECRET_PAGE_URL = "https://octavius.mail.ru/bundles/page.qa.html"
+
+
+STATES = {
+    LOGIN_PAGE_URL: LOGIN_PAGE
+}
