+++ src/page_objects/__init__.py
@@ -1,5 +1,3 @@
 from .page_object import PageObject
 from .consts import LOGIN_PAGE_URL, SECRET_PAGE_URL
 from .filters_applying import FiltersApplyingPageObject
-from .init_page import InitPage
-from .header_object import HeaderObject
